[Rearrange] with-server-component
