+++ apps/web/app/environments/[environmentId]/surveys/PreviewSurvey.tsx
@@ -3,12 +3,13 @@ import Modal from "@/components/preview/Modal";
 import Progress from "@/components/preview/Progress";
 import QuestionConditional from "@/components/preview/QuestionConditional";
 import ThankYouCard from "@/components/preview/ThankYouCard";
-import { useEnvironment } from "@/lib/environments/environments";
-import { useProduct } from "@/lib/products/products";
+// import { useEnvironment } from "@/lib/environments/environments";
+// import { useProduct } from "@/lib/products/products";
 import type { Logic, Question } from "@formbricks/types/questions";
 import { Survey } from "@formbricks/types/surveys";
 import { useEffect, useRef, useState } from "react";
-
+import { TProductWithEnvironmentIds } from "@formbricks/types/v1/product";
+import { TEnvironmentProduct } from "@formbricks/types/v1/environment";
 interface PreviewSurveyProps {
   setActiveQuestionId: (id: string | null) => void;
   activeQuestionId?: string | null;
@@ -19,6 +20,8 @@ interface PreviewSurveyProps {
   thankYouCard: Survey["thankYouCard"];
   autoClose: Survey["autoClose"];
   previewType?: "modal" | "fullwidth" | "email";
+  product: TProductWithEnvironmentIds;
+  environment: TEnvironmentProduct;
 }
 
 export default function PreviewSurvey({
@@ -26,15 +29,14 @@ export default function PreviewSurvey({
   activeQuestionId,
   questions,
   brandColor,
-  environmentId,
+  // environmentId,
   surveyType,
   thankYouCard,
   autoClose,
   previewType,
+  product,
+  environment,
 }: PreviewSurveyProps) {
-  const { environment } = useEnvironment(environmentId);
-  const { product } = useProduct(environmentId);
-
   const [isModalOpen, setIsModalOpen] = useState(true);
   const [progress, setProgress] = useState(0); // [0, 1]
   const [widgetSetupCompleted, setWidgetSetupCompleted] = useState(false);
